Version 1.1.05
Small bug in SparseStumpAlgorithm solved.

Initial commit of SigmoidSingleStumpLearner.

First version of SigmoidSingleStumpLearner using stochastic gradient descent algorithm.

Some type conversion warning had been eliminated.

Small bug in SingleSparseStumpLearner fixed. (It returns NAN if there is no cut at all.)

SigmoidSingleStumpLearner had been implemented. The optimization can be carry out using Stochastic or Batch Gradient Descent.

gradient update fixed in SigmoidSingleStumpLearner

SigmoidSingleStump bug fixed (bgd version)

only bracketing

SigmoidSingleStumpLearner's update rules are simplified

Target function in SigmoidStumpLearner can be chosen.

Small bug is fixed in SigmoidSingleStump

online learning implementation starting

online learning implementation in progress

FilterBoost can train the base learner online.
The ConstantLearner can be trained online.
SigmoidSingleStumpLearner can be trained online.

The SigmoidSingleStumpLearner's phi function is updated. Now, it returns {-1,+1} and the _v vector is also in {-1,1}^K

LinearCombinationLearner is added.

FilterBoost is cleaned up.

LinearCombinationLearner is added.

LogisticRegressionLearner's implementation has started.

The edge estimate is after the stochastic update.

The decreasing of sigmoidsinglestumplearner can be set in command line.

AdaLine implementation is started.
LeinarCombinationLearner is renamed to NeuronLearner.

AdaLineLEarner v1.0

Stochastic Learneren.cpp added

subCopyState is fixed in constant learner.

Updated the StochasticLearner interface.

Fix bugs with svmlight parser ("qid" header and comments after the features/values enumeration) please remember the format is specified as this :
    <line> .=. <target> <feature>:<value> <feature>:<value> ... <feature>:<value> # <info>
    <target> .=. +1 | -1 | 0 | <float>
    <feature> .=. <integer> | "qid"
    <value> .=. <float>
    <info> .=. <string>

Minor bug: the verbosity of the output when using --posteriors was always equal to 1.

Minor bug fix in the margin calculation.

Missing semicolon in OutputInfo.cpp

Modified the bibtex of the main page in doxygen.

Example.h reindented

getLabelY(const int labelIdx) and  getLabelWeight(const int labelIdx) is added to Example. Put detailed comments for the functions, pliiiz we should _always_ comment new functions.

getLabel(const int labelIdx) is added. It returns a static empty Label if the label is not enumerated. A constructor for Label is also added.

getLabelY and getLabelWeight are const functions now.

SoftCascadeLearner bug is fixed with the use of the new getLabelY function.

SoftCascadeClassifier now uses the getLabelY function.

VJCascadeLearner bug is fixed with the use of the new getLabelY function.

getLabelY replacement inside SoftCascadeOutput::computeAndOutput of OutputInfo

Reindentation commit, this time for real. no-tabs and 4-spaces.

replaced make_pair<> templates (which are incorrect) with pair<>

Bug fixes in OutputInfo:
- Header minor fix.
- Minimum margins calculation fix.
- Droped the negmargin output.

cherry-picking ArcGV from Stochastic branch

+++ src/IO/InputData.cpp
@@ -51,8 +51,8 @@
 
 //
 namespace MultiBoost {
-	
-	// ------------------------------------------------------------------------
+        
+    // ------------------------------------------------------------------------
     
     
     void InputData::addExample( Example example ) { 
@@ -79,67 +79,67 @@ namespace MultiBoost {
         
         //otherwise : _subsetAlreadyComputed = false;
     }
-	
+        
     // ------------------------------------------------------------------------
-	
-	int	InputData::loadIndexSet( set< int > ind ) {
-		int i = 0;
-		//upload the indirection
-		for (int j = 0; j < this->_rawIndices.size(); j++) this->_rawIndices[j] = -1;
+        
+    int     InputData::loadIndexSet( set< int > ind ) {
+        int i = 0;
+        //upload the indirection
+        for (int j = 0; j < this->_rawIndices.size(); j++) this->_rawIndices[j] = -1;
 
-		map<int, int> tmpPointsPerClass;
-		
-		for( set< int >::iterator it = ind.begin(); it != ind.end(); it++ ) {
-			this->_indirectIndices[i] = *it;
-			this->_rawIndices[*it] = i;
-			
-			i++;
-			
-			const vector<Label>& labels = _pData->getLabels( *it );
-			vector<Label>::const_iterator lIt;
-			
-			for (lIt = labels.begin(); lIt != labels.end(); ++lIt )
-			{
-				switch ( this->_pData->getLabelRep() ) {
-					case LR_DENSE:
-						if ( lIt->y > 0 )
-							tmpPointsPerClass[lIt->idx]++;
-						break;
-					case LR_SPARSE:
-						if ( lIt->y > 0 )
-							tmpPointsPerClass[lIt->idx]++;
-						break;
-				}
-			}
-		}
-		
-		_nExamplesPerClass.clear();
-		for (int l = 0; l < this->_pData->getNumClasses(); ++l)
-			_nExamplesPerClass.push_back( tmpPointsPerClass[l] );
-		
-		
-		this->_numExamples = ind.size();
+        map<int, int> tmpPointsPerClass;
+                
+        for( set< int >::iterator it = ind.begin(); it != ind.end(); it++ ) {
+            this->_indirectIndices[i] = *it;
+            this->_rawIndices[*it] = i;
+                        
+            i++;
+                        
+            const vector<Label>& labels = _pData->getLabels( *it );
+            vector<Label>::const_iterator lIt;
+                        
+            for (lIt = labels.begin(); lIt != labels.end(); ++lIt )
+            {
+                switch ( this->_pData->getLabelRep() ) {
+                case LR_DENSE:
+                    if ( lIt->y > 0 )
+                        tmpPointsPerClass[lIt->idx]++;
+                    break;
+                case LR_SPARSE:
+                    if ( lIt->y > 0 )
+                        tmpPointsPerClass[lIt->idx]++;
+                    break;
+                }
+            }
+        }
+                
+        _nExamplesPerClass.clear();
+        for (int l = 0; l < this->_pData->getNumClasses(); ++l)
+            _nExamplesPerClass.push_back( tmpPointsPerClass[l] );
+                
+                
+        this->_numExamples = (int)ind.size();
         
         _subsetAlreadyComputed = false;
         
-		return 0;
-	}
-	
-	
-	// ------------------------------------------------------------------------
-	
-	void InputData::clearIndexSet( void ) {
-		
-		for( int i = 0; i < this->_pData->getNumExample(); i++ ) {
-			this->_indirectIndices[ i ] = i;
-			this->_rawIndices[ i ] = i;
-		}
-		this->_numExamples = this->_pData->getNumExample();
-		
-		_nExamplesPerClass.clear();
-		_nExamplesPerClass = this->_pData->getExamplesPerClass();
+        return 0;
+    }
+        
+        
+    // ------------------------------------------------------------------------
+        
+    void InputData::clearIndexSet( void ) {
+                
+        for( int i = 0; i < this->_pData->getNumExample(); i++ ) {
+            this->_indirectIndices[ i ] = i;
+            this->_rawIndices[ i ] = i;
+        }
+        this->_numExamples = this->_pData->getNumExample();
+                
+        _nExamplesPerClass.clear();
+        _nExamplesPerClass = this->_pData->getExamplesPerClass();
         
         _subsetAlreadyComputed = false;
-	}
-	
+    }
+        
 } // end of namespace MultiBoost
